Introduced a corpus module
