+++ src/iranlowo/preprocessing.py
@@ -1,5 +1,6 @@
 import csv
 import gzip
+import unicodedata
 from pathlib import Path
 
 
@@ -89,4 +90,131 @@ def convert_to_owé_format(path, sep=4, outpath=None, to_csv=False, **kwargs):
                     writer.write(en_text)
         except IndexError:
             pass  # End of file reached
+
+
+def strip_accents_text(text_string):
+    """
+    Converts the string to NFD, separates & returns only the base characters
+    :param text_string:
+    :return: input string without diacritic adornments on base characters
+    """
+    return "".join(
+        c
+        for c in unicodedata.normalize("NFD", text_string)
+        if unicodedata.category(c) != "Mn"
+    )
+
+
+def strip_accents_file(filename, outfilename):
+    """
+    Reads filename containing diacritics, converts to NFC for consistency,
+    then writes outfilename with diacritics removed
+    :param filename:
+    :param outfilename:
+    :return: None
+    """
+    text = "".join(
+        c for c in unicodedata.normalize("NFC", open(filename, encoding="utf-8").read())
+    )
+    try:
+        f = open(outfilename, "w")
+    except EnvironmentError:
+        return False
+    else:
+        with f:
+            f.write(strip_accents_text(text))
+        return True
+
+
+def normalize_diacritics_text(text_string):
+    """Convenience wrapper to abstract away unicode & NFC"""
+    return unicodedata.normalize("NFC", text_string)
+
+
+def normalize_diacritics_file(filename, outfilename):
+    """File based Convenience wrapper to abstract away unicode & NFC"""
+    try:
+        text = "".join(
+            c
+            for c in unicodedata.normalize(
+                "NFC", open(filename, encoding="utf-8").read()
+            )
+        )
+        with open(outfilename, "w", encoding="utf-8") as f:
+            f.write(text)
+    except EnvironmentError:
+        return False
+    else:
+        return True
+
+
+def split_corpus_on_symbol(filename, outfilename, symbol=","):
+    """
+    For yoruba blog (and probably bibeli mimo)
+
+    Args: filenames for I/O and symbol to split lines on
+    Returns: writes outputfile
+    :param filename: input file
+    :param outfilename: processed output file to write
+    :param symbol: to split lines on
+    :return: None, with side-effect of writing an outputfile
+    """
+
+    lines = tuple(open(filename, "r", encoding="utf-8"))
+
+    min_words_to_split = 10
+    min_words_in_utt = 5
+
+    with open(outfilename, "w") as f:
+        # split out heavily comma'd text :((
+        for line in lines:
+            if symbol in line:
+                num_words = len(line.split())
+                num_commas = line.count(symbol)
+                curr_comma_position = line.index(symbol)
+                num_words_ahead_of_curr_comma = len(line[0:curr_comma_position].split())
+
+                curr_line = line
+                while num_commas > 0:
+                    if num_words < min_words_to_split:
+                        # print(curr_line.strip())
+                        f.write(curr_line)
+                        break
+                    if num_words >= min_words_to_split:
+                        if (
+                                num_words_ahead_of_curr_comma >= min_words_in_utt
+                                and len(curr_line[curr_comma_position:].split())
+                                >= min_words_in_utt
+                        ):
+                            f.write(curr_line[0:curr_comma_position] + "\n")
+
+                            # update vars
+                            curr_line = curr_line[curr_comma_position + 1:]
+                            num_words = len(curr_line.split())
+                            num_commas = num_commas - 1
+                            if num_commas > 0:
+                                curr_comma_position = curr_line.index(symbol)
+                                num_words_ahead_of_curr_comma = len(
+                                    curr_line[0:curr_comma_position].split()
+                                )
+                            else:
+                                f.write(curr_line)
+                        else:
+                            # ignore too short comma (+= vs = on current comma position)
+                            num_commas = num_commas - 1
+                            if num_commas > 0:  # for say 3 commas
+                                curr_comma_position += (
+                                        curr_line[curr_comma_position + 1:].index(symbol)
+                                        + 1
+                                )
+                                num_words_ahead_of_curr_comma = len(
+                                    curr_line[0:curr_comma_position].split()
+                                )
+                            else:
+                                f.write(curr_line)
+                    else:
+                        f.write(curr_line)
+            else:
+                f.write(line)
+
